+++ app/src/main/java/info/appdev/chartexample/RealtimeLineChartActivity.java
@@ -113,7 +113,7 @@ public class RealtimeLineChartActivity extends DemoBase implements
             }
 
             int cycleValue = (int) (set.getEntryCount() % 100.0);
-            data.addEntry(new Entry(set.getEntryCount(), (float) (sampleValues[cycleValue].floatValue() * 40) + 30f), 0);
+            data.addEntry(new Entry(set.getEntryCount(), (sampleValues[cycleValue].floatValue() * 40) + 30f), 0);
             data.notifyDataChanged();
 
             // let the chart know it's data has changed
